No modules for sources under test in the ugliest possible way
This will need some better fixing. Essentially we don't want to expose
modulemaps of code that is under test to the test target, as otherwise
we won't be able to import any private headers (or public headers with
quote import syntax, as they'll complain about not being modular).

This is a really, really bad hack that changes the location where the
private test headers come from, so that it can override the headers
style of the library under test. Need to figure out a better way of
doing this, current ideas:

- Be able to pass down what kind of header you want like CxxPlatform.
  This would be ideal but requires a change in /a lot/ of places.
- Not use `module.modulemap` but something else so modules aren't
  implicitly found. This would require a `-fmodule-map-file` for every
  modulemap everywhere, in addition to the header search path.
- Have modulemaps anyway and make users change their imports?